separate repodata logic from index into a new core/repodata.py module
