+++ tests/core/test_index.py
@@ -11,8 +11,9 @@ from conda.common.io import env_var
 from conda.base.context import context, reset_context
 from conda.common.compat import iteritems
 from conda.common.disk import temporary_content_in_file
-from conda.core.index import get_index, read_mod_and_etag, cache_fn_url, \
+from conda.core.index import get_index, cache_fn_url, \
     Response304ContentUnchanged
+from conda.core.repodata import read_mod_and_etag, Response304ContentUnchanged, cache_fn_url
 
 try:
     from unittest.mock import patch
