+++ tests/test_create.py
@@ -43,7 +43,7 @@ from conda.common.io import captured, disable_logger, replace_log_streams, stder
 from conda.common.path import get_bin_directory_short_path, get_python_site_packages_short_path, pyc_path
 from conda.common.url import path_to_url
 from conda.common.yaml import yaml_load
-from conda.core.index import create_cache_dir
+from conda.core.repodata import create_cache_dir
 from conda.core.linked_data import get_python_version_for_prefix, \
     linked as install_linked, linked_data, linked_data_
 from conda.core.package_cache import PackageCache
